+++ app/helpers/admin/claims_helper.rb
@@ -103,6 +103,14 @@ module Admin
       end
     end
 
+    def payroll_run_status(claim)
+      if claim.payrolled?
+        link_to(claim.payment.payroll_run.created_at.strftime("%B %Y"), admin_payroll_run_path(claim.payment.payroll_run))
+      else
+        "Awaiting payroll"
+      end
+    end
+
     private
 
     def matching_attributes_for(claim)